requirements.txt
django==3.1.5
requests==2.25.1
xmltodict==0.12.0
graphene-django==2.15.0
# psycopg2==2.8.6
